+++ src/r0/account.rs
@@ -8,6 +8,7 @@ pub mod get_username_availability;
 pub mod register;
 pub mod request_3pid_management_token_via_email;
 pub mod request_3pid_management_token_via_msisdn;
+pub mod request_openid_token;
 pub mod request_password_change_token_via_email;
 pub mod request_password_change_token_via_msisdn;
 pub mod request_registration_token_via_email;